Update test
When solving for a single parameter, diff in solution to Newton system due to
fact that 'inner' does not behave exactly the same in Vl*Vl and Vl.
The relative difference in search direction can be large (~1e-1), and I was not
able to fix this (tried rescaling vectors before applying inner).
In the end, it seems to have very little impact on the quality of the search
direction, and it is probably not worth investing more time into that.

If something had to be done, I would modify' mult' in objectiveacoustic.py to
accept vectors in Vl when inverting for a single param. This would also require
modifying 'init_vector' for compatibility with CGSolverSteihaug.py. This just
feels to be the right thing to do, but again, this does not appear critical.

No more action taken at the moment.